+++ syncstorage/src/tokenserver/migrations/2021-08-03-234845_populate_services/up.sql
@@ -1,0 +1,3 @@
+INSERT IGNORE INTO services (id, service, pattern) VALUES
+    (1, "sync-1.1", "{node}/1.1/{uid}"),
+    (2, "sync-1.5", "{node}/1.5/{uid}");